# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AppApiKeyResponse": {
        "description": "The current (reveal) or freshly-minted (rotate) app ``X-Api-Key``, in plaintext.\n\nAuthenticated-only (Plex session or currently-valid ``X-Api-Key``): reveal is\nthe belt-and-braces recovery path for a lost/forgotten key on a device that\nstill has it saved, and rotate mints and returns a brand-new key ONCE \u2014 the\nplaintext is never retrievable again after this response, only the\nFernet-encrypted column at rest (matching the one-time disclosure setup's\n``/complete`` already gives the initial key).",
        "properties": {
          "app_api_key": {
            "title": "App Api Key",
            "type": "string"
          }
        },
        "required": [
          "app_api_key"
        ],
        "title": "AppApiKeyResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlexLibraryOption": {
        "description": "One movie- OR tv-library folder Plex reports, with whether the app can write to it.\n\n``path`` is a Plex library location (from Plex's own ``/library/sections``), so\nchoosing it for ``movies_root`` / ``tv_root`` avoids a typed path entirely (and\nthe path\u2194section mismatch that breaks a targeted scan). ``section_type`` tags\nwhich root this option is for (``\"movie\"`` -> ``movies_root``, ``\"tv\"`` ->\n``tv_root``) so a single generalized list can drive both pickers. ``writable``\nis the app's own check (``None`` when not probed \u2014 see the field): a\nknown-not-writable location is the split-mount signal \u2014 surfaced, not hidden.",
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          },
          "section_key": {
            "title": "Section Key",
            "type": "string"
          },
          "section_type": {
            "enum": [
              "movie",
              "tv"
            ],
            "title": "Section Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "writable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Writable"
          }
        },
        "required": [
          "section_key",
          "title",
          "path",
          "section_type"
        ],
        "title": "PlexLibraryOption",
        "type": "object"
      },
...
      "SettingsResponse": {
        "description": "Redacted view of stored service config.\n\nPlaintext (non-secret) values are returned as-is; secret values are shown as\n``\"***\"`` when configured and ``None`` when unset \u2014 the plaintext secret is\nNEVER serialized.",
        "properties": {
          "anime_movie_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anime Movie Root"
          },
          "anime_tv_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anime Tv Root"
          },
          "auto_grab_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Grab Enabled"
          },
          "disk_pressure_target_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disk Pressure Target Percent"
          },
          "disk_pressure_threshold_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disk Pressure Threshold Percent"
          },
          "eviction_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eviction Enabled"
          },
          "eviction_grace_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eviction Grace Days"
          },
          "eviction_interval_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eviction Interval Minutes"
          },
          "eviction_proactive_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eviction Proactive Enabled"
          },
          "log_retention_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Retention Days"
          },
          "movies_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movies Root"
          },
          "plex_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plex Token"
          },
          "plex_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plex Url"
          },
          "prowlarr_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prowlarr Api Key"
          },
          "prowlarr_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prowlarr Url"
          },
          "qbittorrent_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qbittorrent Password"
          },
          "qbittorrent_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qbittorrent Url"
          },
          "qbittorrent_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qbittorrent Username"
          },
          "tmdb_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Api Key"
          },
          "tv_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tv Root"
          }
        },
        "title": "SettingsResponse",
        "type": "object"
      },
      "SettingsUpdate": {
        "description": "Partial upsert of service config (``PUT /settings``).\n\nOnly fields present in the request body are written; secret fields are stored\nencrypted at rest. ``None`` / absent fields are left unchanged.",
        "properties": {
          "anime_movie_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anime Movie Root"
          },
          "anime_tv_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anime Tv Root"
          },
          "auto_grab_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Grab Enabled"
          },
          "disk_pressure_target_percent": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disk Pressure Target Percent"
          },
          "disk_pressure_threshold_percent": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disk Pressure Threshold Percent"
          },
          "eviction_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eviction Enabled"
          },
          "eviction_grace_days": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eviction Grace Days"
          },
          "eviction_interval_minutes": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eviction Interval Minutes"
          },
          "eviction_proactive_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eviction Proactive Enabled"
          },
          "log_retention_days": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Retention Days"
          },
          "movies_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movies Root"
          },
          "plex_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plex Token"
          },
          "plex_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plex Url"
          },
          "prowlarr_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prowlarr Api Key"
          },
          "prowlarr_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prowlarr Url"
          },
          "qbittorrent_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qbittorrent Password"
          },
          "qbittorrent_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qbittorrent Url"
          },
          "qbittorrent_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qbittorrent Username"
          },
          "tmdb_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Api Key"
          },
          "tv_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tv Root"
          }
        },
        "title": "SettingsUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Plex Manager",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "description": "Return the redacted service config (secrets shown as ``\"***\"``).",
        "operationId": "get_settings_endpoint_api_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Settings Endpoint",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "description": "Upsert the provided config and return the redacted result.\n\nA secret field whose incoming value is the redaction mask (``\"***\"``) is a\nno-op: GET returns ``\"***\"`` for a configured secret, so a FE that round-trips\nthe whole object back (e.g. after editing only ``plex_url``) must not clobber\nthe real credential with the mask \u2014 a silent secret-wipe that would only\nsurface later as an auth failure to the downstream service.\n\nThe disk-pressure pair is cross-checked against the EFFECTIVE (post-update)\nvalues, not just what this one request happens to carry: ``SettingsUpdate``'s\nown ``model_validator`` only catches a target above the threshold when BOTH\nare submitted together, but ``PUT`` is a PARTIAL update \u2014 sending just one\nside against an already-stored (and now-inverted) other side would otherwise\nsilently leave the whole threshold-to-target band unable to relieve pressure\n(see :func:`~plex_manager.web.routers.settings._validate_disk_pressure_pair`).\nChecked, and rejected with the SAME 422 shape, BEFORE anything is written.",
        "operationId": "put_settings_endpoint_api_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Put Settings Endpoint",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/app-key": {
      "get": {
        "description": "Return the current app ``X-Api-Key`` in plaintext.\n\nAuthenticated: the caller already proved they have a valid Plex session or\napp key, so this is not an anonymous disclosure -- it is the break-glass\nrecovery path for a NEW device/browser that needs to be paired without\nre-running setup.",
        "operationId": "reveal_app_key_endpoint_api_v1_settings_app_key_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppApiKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Reveal App Key Endpoint",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/app-key/rotate": {
      "post": {
        "description": "Mint a brand-new app ``X-Api-Key``, invalidating the old one, and return it once.\n\nEvery OTHER device/browser with the OLD key saved (localStorage) is\nimmediately locked out -- there is exactly one live key at a time, matching\n``require_api_key``'s single-key comparison. The frontend persists the\nreturned key immediately for future access-key recovery, but normal browser\nauth continues to use the Plex session cookie.\n\nCompare-and-swap against concurrent rotations: two rotate requests carrying\nthe SAME old key can both clear ``require_api_key`` (each reads the old stored\nvalue) before either commits. Without a guard the write that commits second\nwould silently overwrite the first's freshly minted key, so the client that\nfired the first request would be left displaying an already-dead key. The\nre-read/compare/mint/commit is run under the module-level ``_rotate_lock`` so it\nis a true atomic read-modify-write rather than check-then-act: the compare and\nthe write cannot interleave with another rotation, so the loser's re-read runs\nonly AFTER the winner has committed. Inside THIS request's own transaction we\nre-read the stored key and require it to still equal the key the request\nauthenticated with; if it has already changed, the race happened and we answer\n409 (``app_key_changed``) rather than clobber the winner. The check is skipped\nunder ``dev_auth_bypass`` (there is no authenticated key to compare against),\nexactly like ``require_api_key`` itself.",
        "operationId": "rotate_app_key_endpoint_api_v1_settings_app_key_rotate_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppApiKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Rotate App Key Endpoint",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/plex-libraries": {
      "get": {
        "description": "Library folders (movie AND tv) Plex reports, for the Settings\n``movies_root`` / ``tv_root`` pickers -- each option is tagged by\n``section_type`` so the frontend can filter to the picker it's rendering.\n\nUses the stored Plex creds (no re-typing the token); 409 if Plex is unconfigured.",
        "operationId": "plex_libraries_endpoint_api_v1_settings_plex_libraries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlexLibraryOption"
                  },
                  "title": "Response Plex Libraries Endpoint Api V1 Settings Plex Libraries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Plex Libraries Endpoint",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}